backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sentence-transformers==2.2.2
scikit-learn==1.4.0
numpy==1.26.3
google-generativeai==0.5.0
python-multipart==0.0.6
pydantic==2.5.3
